+++ src/renderer/utils/constants.ts
@@ -1,0 +1,1 @@
+export const STANDALONE_PROJECT_ID = '__standalone__'